requirements.txt
greenlet==3.0.1
joblib==1.3.2
mysql-connector-python==8.2.0
numpy==1.26.2
pandas==2.1.3
protobuf==4.21.12
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.2
scikit-surprise==1.1.3
scipy==1.11.4
six==1.16.0
SQLAlchemy==2.0.23
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
